feat-143: 검색 화면 검색 결과 화면 통합
