+++ presentation/src/main/java/com/plub/presentation/ui/main/home/search/SearchingFragment.kt
@@ -6,9 +6,20 @@ import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
+import com.google.android.material.tabs.TabLayout
+import com.plub.domain.model.enums.DialogMenuItemType
+import com.plub.domain.model.enums.DialogMenuType
+import com.plub.domain.model.enums.PlubCardType
+import com.plub.domain.model.enums.PlubSearchType
+import com.plub.domain.model.enums.PlubSortType
+import com.plub.presentation.R
 import com.plub.presentation.base.BaseFragment
 import com.plub.presentation.databinding.FragmentSearchingBinding
+import com.plub.presentation.databinding.IncludeTabSearchResultBinding
 import com.plub.presentation.ui.common.decoration.GridSpaceDecoration
+import com.plub.presentation.ui.common.dialog.SelectMenuBottomSheetDialog
+import com.plub.presentation.ui.common.dialog.adapter.DialogMenuAdapter
+import com.plub.presentation.ui.main.home.adapter.PlubCardAdapter
 import com.plub.presentation.ui.main.home.search.adapter.RecentSearchAdapter
 import com.plub.presentation.util.px
 import dagger.hilt.android.AndroidEntryPoint
@@ -20,6 +31,7 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
 ) {
 
     companion object {
+        private const val POSITION_TOP = 0
         private const val TOTAL_SPAN_SIZE = 2
         private const val ITEM_SPAN_SIZE = 1
         private const val ITEM_VERTICAL_SPACE = 8
@@ -28,14 +40,26 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
 
     override val viewModel: SearchingViewModel by viewModels()
 
-    private val listAdapter: RecentSearchAdapter by lazy {
+    private val recentListAdapter: RecentSearchAdapter by lazy {
         RecentSearchAdapter(object : RecentSearchAdapter.Delegate {
             override fun onClickDelete(text: String) {
                 viewModel.onDeleteRecentSearch(text)
             }
 
             override fun onClickRecentSearch(text: String) {
-                viewModel.onRecentSearch(text)
+                viewModel.onClickRecentSearch(text)
+            }
+        })
+    }
+
+    private val searchListAdapter: PlubCardAdapter by lazy {
+        PlubCardAdapter(object : PlubCardAdapter.Delegate {
+            override fun onClickBookmark(id: Int) {
+                viewModel.onClickBookmark(id)
+            }
+
+            override fun onClickPlubCard(id: Int) {
+
             }
         })
     }
@@ -43,6 +67,8 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
     override fun initView() {
         binding.apply {
             vm = viewModel
+
+            initTab()
 
             recyclerViewRecentSearch.apply {
                 layoutManager = GridLayoutManager(context, TOTAL_SPAN_SIZE).apply {
@@ -60,13 +86,27 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
                         false
                     )
                 )
-                adapter = listAdapter
+                adapter = recentListAdapter
+            }
+
+            recyclerViewSearch.apply {
+                layoutManager = GridLayoutManager(context, PlubCardType.TOTAL_SPAN_SIZE).apply {
+                    spanSizeLookup = object : GridLayoutManager.SpanSizeLookup() {
+                        override fun getSpanSize(position: Int): Int {
+                            val viewType = searchListAdapter.getItemViewType(position)
+                            val cardType = PlubCardType.valueOf(viewType)
+                            return cardType.spanSize
+                        }
+                    }
+                }
+
+                adapter = searchListAdapter
             }
 
             editTextSearch.apply {
                 setOnEditorActionListener { _, keyCode, keyEvent ->
                     if(keyCode == EditorInfo.IME_ACTION_SEARCH) {
-                        viewModel.onSearch(text.toString())
+                        viewModel.onSearchPlubRecruit(text.toString())
                         true
                     }else false
                 }
@@ -82,7 +122,9 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
         repeatOnStarted(viewLifecycleOwner) {
             launch {
                 viewModel.uiState.collect {
-                    listAdapter.submitList(it.recentSearchList)
+                    recentListAdapter.submitList(it.recentSearchList)
+                    searchListAdapter.submitList(it.searchList)
+                    setSortTypeText(it.sortType)
                 }
             }
 
@@ -94,26 +136,76 @@ class SearchingFragment : BaseFragment<FragmentSearchingBinding, SearchingPageSt
         }
     }
 
+    private fun initTab() {
+        binding.tabLayoutSearchType.apply {
+            PlubSearchType.values().forEach {
+                addTab(getTabView(it))
+            }
+            addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
+                override fun onTabSelected(tab: TabLayout.Tab) {
+                    viewModel.onTabSelected(tab.position)
+                }
+
+                override fun onTabReselected(tab: TabLayout.Tab) {}
+                override fun onTabUnselected(tab: TabLayout.Tab) {}
+            })
+        }
+    }
+
+    private fun getTabView(searchType: PlubSearchType): TabLayout.Tab {
+        val tab = binding.tabLayoutSearchType.newTab()
+        return tab.apply {
+            customView = IncludeTabSearchResultBinding.inflate(layoutInflater).apply {
+                textViewTab.text = getTabTitle(searchType)
+            }.root
+        }
+    }
+
+    private fun getTabTitle(searchType: PlubSearchType): String {
+        return when (searchType) {
+            PlubSearchType.TITLE -> getString(R.string.search_result_search_type_title)
+            PlubSearchType.NAME -> getString(R.string.search_result_search_type_name)
+            PlubSearchType.TITLE_AND_CONTENTS -> getString(R.string.search_result_search_type_title_and_content)
+        }
+    }
+
+    private fun inspectEventFlow(event: SearchingEvent) {
+        when (event) {
+            is SearchingEvent.HideKeyboard -> {
+                hideKeyboard()
+                binding.editTextSearch.clearFocus()
+            }
+            is SearchingEvent.ScrollToTop -> scrollToTop()
+            is SearchingEvent.ShowSelectSortTypeBottomSheetDialog -> showSelectSortTypeDialog(event.selectedItem)
+        }
+    }
+
+    private fun setSortTypeText(sortType: PlubSortType) {
+        val sortTypeRes = when (sortType) {
+            PlubSortType.POPULAR -> R.string.word_sort_type_popular
+            PlubSortType.NEW -> R.string.word_sort_type_new
+        }
+        binding.textViewSortType.text = getString(sortTypeRes)
+    }
+
+    private fun scrollToTop() {
+        binding.recyclerViewSearch.scrollToPosition(POSITION_TOP)
+    }
+
     private fun hideKeyboard() {
         val imm = context?.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.hideSoftInputFromWindow(binding.editTextSearch.windowToken, 0)
     }
 
-    private fun clear() {
-        binding.editTextSearch.setText("")
-        binding.editTextSearch.clearFocus()
-    }
-
-    private fun goToSearchResult(search:String) {
-        val action = SearchingFragmentDirections.actionSearchingToSearchResult(search)
-        findNavController().navigate(action)
-    }
-
-    private fun inspectEventFlow(event: SearchingEvent) {
-        when (event) {
-            is SearchingEvent.GoToSearchResult -> goToSearchResult(event.search)
-            is SearchingEvent.HideKeyboard -> hideKeyboard()
-            is SearchingEvent.Clear -> clear()
-        }
+    private fun showSelectSortTypeDialog(selectedMenuType: DialogMenuItemType) {
+        val title = getString(R.string.word_sort)
+        SelectMenuBottomSheetDialog.newInstance(
+            DialogMenuType.SORT_TYPE,
+            title,
+            DialogMenuAdapter.VIEW_TYPE_SPINNER,
+            selectedMenuType
+        ) {
+            viewModel.onClickSortMenuItemType(it)
+        }.show(parentFragmentManager, "")
     }
 }
