--- presentation/src/main/java/com/plub/presentation/ui/main/home/searchResult/SearchResultEvent.kt
@@ -1,11 +1,0 @@
-package com.plub.presentation.ui.main.home.searchResult
-
-import com.plub.domain.model.enums.DialogMenuItemType
-import com.plub.presentation.ui.Event
-
-sealed class SearchResultEvent : Event {
-    object HideKeyboard : SearchResultEvent()
-    object ScrollToTop : SearchResultEvent()
-    object NavigationPopEvent : SearchResultEvent()
-    data class ShowSelectSortTypeBottomSheetDialog(val selectedItem: DialogMenuItemType) : SearchResultEvent()
-}
